Support lifecycle hooks for CloneSet (#362)

+++ pkg/controller/cloneset/scale/cloneset_scale.go
@@ -11,6 +11,7 @@ import (
 	clonesetutils "github.com/openkruise/kruise/pkg/controller/cloneset/utils"
 	"github.com/openkruise/kruise/pkg/util"
 	"github.com/openkruise/kruise/pkg/util/expectations"
+	"github.com/openkruise/kruise/pkg/util/lifecycle"
 	apps "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
@@ -37,14 +38,13 @@ type Interface interface {
 }
 
 // New returns a scale control.
-func New(c client.Client, recorder record.EventRecorder, exp expectations.ScaleExpectations) Interface {
-	return &realControl{Client: c, recorder: recorder, exp: exp}
+func New(c client.Client, recorder record.EventRecorder) Interface {
+	return &realControl{Client: c, recorder: recorder}
 }
 
 type realControl struct {
 	client.Client
 	recorder record.EventRecorder
-	exp      expectations.ScaleExpectations
 }
 
 func (r *realControl) Manage(
@@ -63,13 +63,22 @@ func (r *realControl) Manage(
 		return false, nil
 	}
 
-	if podsToDelete := getPodsToDelete(updateCS, pods); len(podsToDelete) > 0 {
-		klog.V(3).Infof("CloneSet %s begin to delete pods in podsToDelete: %v", controllerKey, podsToDelete)
-		return r.deletePods(updateCS, podsToDelete, pvcs)
+	podsSpecifiedToDelete, podsInPreDelete := getPlannedDeletedPods(updateCS, pods)
+	podsToDelete := util.MergePods(podsSpecifiedToDelete, podsInPreDelete)
+	if len(podsToDelete) > 0 {
+		klog.V(3).Infof("CloneSet %s find pods %v specified to delete and pods %v in preDelete",
+			controllerKey, util.GetPodNames(podsSpecifiedToDelete).List(), util.GetPodNames(podsInPreDelete).List())
+
+		if modified, err := r.managePreparingDelete(updateCS, pods, podsInPreDelete, len(podsToDelete)); err != nil || modified {
+			return modified, err
+		}
+
+		if modified, err := r.deletePods(updateCS, podsToDelete, pvcs); err != nil || modified {
+			return modified, err
+		}
 	}
 
 	updatedPods, notUpdatedPods := clonesetutils.SplitPodsByRevision(pods, updateRevision)
-
 	diff, currentRevDiff := calculateDiffs(updateCS, updateRevision == currentRevision, len(pods), len(notUpdatedPods))
 
 	if diff < 0 {
@@ -96,6 +105,11 @@ func (r *realControl) Manage(
 			currentCS, updateCS, currentRevision, updateRevision, availableIDs.List(), existingPVCNames)
 
 	} else if diff > 0 {
+		if len(podsToDelete) > 0 {
+			klog.V(3).Infof("CloneSet %s skip to scale in %d for existing pods to delete", controllerKey, diff)
+			return false, nil
+		}
+
 		klog.V(3).Infof("CloneSet %s begin to scale in %d pods including %d (current rev)",
 			controllerKey, diff, currentRevDiff)
 
@@ -105,6 +119,30 @@ func (r *realControl) Manage(
 	}
 
 	return false, nil
+}
+
+func (r *realControl) managePreparingDelete(cs *appsv1alpha1.CloneSet, pods, podsInPreDelete []*v1.Pod, numToDelete int) (bool, error) {
+	diff := int(*cs.Spec.Replicas) - len(pods) + numToDelete
+	var modified bool
+	for _, pod := range podsInPreDelete {
+		if diff <= 0 {
+			return modified, nil
+		}
+		if isPodSpecifiedDelete(cs, pod) {
+			continue
+		}
+
+		klog.V(3).Infof("CloneSet %s patch pod %s lifecycle from PreparingDelete to Normal",
+			clonesetutils.GetControllerKey(cs), pod.Name)
+		if patched, err := lifecycle.PatchPodLifecycle(r, pod, appsv1alpha1.LifecycleStateNormal); err != nil {
+			return modified, err
+		} else if patched {
+			modified = true
+			clonesetutils.ResourceVersionExpectations.Expect(pod)
+		}
+		diff--
+	}
+	return modified, nil
 }
 
 func (r *realControl) createPods(
@@ -123,7 +161,7 @@ func (r *realControl) createPods(
 
 	podsCreationChan := make(chan *v1.Pod, len(newPods))
 	for _, p := range newPods {
-		r.exp.ExpectScale(clonesetutils.GetControllerKey(updateCS), expectations.Create, p.Name)
+		clonesetutils.ScaleExpectations.ExpectScale(clonesetutils.GetControllerKey(updateCS), expectations.Create, p.Name)
 		podsCreationChan <- p
 	}
 
@@ -136,6 +174,7 @@ func (r *realControl) createPods(
 		if pod.Labels[apps.ControllerRevisionHashLabelKey] == currentRevision {
 			cs = currentCS
 		}
+		lifecycle.SetPodLifecycle(appsv1alpha1.LifecycleStateNormal)(pod)
 
 		var createErr error
 		if createErr = r.createOnePod(cs, pod, existingPVCNames); createErr != nil {
@@ -151,7 +190,7 @@ func (r *realControl) createPods(
 	// rollback to ignore failure pods because the informer won't observe these pods
 	for _, pod := range newPods {
 		if _, ok := successPodNames.Load(pod.Name); !ok {
-			r.exp.ObserveScale(clonesetutils.GetControllerKey(updateCS), expectations.Create, pod.Name)
+			clonesetutils.ScaleExpectations.ObserveScale(clonesetutils.GetControllerKey(updateCS), expectations.Create, pod.Name)
 		}
 	}
 
@@ -167,9 +206,9 @@ func (r *realControl) createOnePod(cs *appsv1alpha1.CloneSet, pod *v1.Pod, exist
 		if existingPVCNames.Has(c.Name) {
 			continue
 		}
-		r.exp.ExpectScale(clonesetutils.GetControllerKey(cs), expectations.Create, c.Name)
+		clonesetutils.ScaleExpectations.ExpectScale(clonesetutils.GetControllerKey(cs), expectations.Create, c.Name)
 		if err := r.Create(context.TODO(), &c); err != nil {
-			r.exp.ObserveScale(clonesetutils.GetControllerKey(cs), expectations.Create, c.Name)
+			clonesetutils.ScaleExpectations.ObserveScale(clonesetutils.GetControllerKey(cs), expectations.Create, c.Name)
 			r.recorder.Eventf(cs, v1.EventTypeWarning, "FailedCreate", "failed to create pvc: %v, pvc: %v", err, util.DumpJSON(c))
 			return err
 		}
@@ -185,15 +224,27 @@ func (r *realControl) createOnePod(cs *appsv1alpha1.CloneSet, pod *v1.Pod, exist
 }
 
 func (r *realControl) deletePods(cs *appsv1alpha1.CloneSet, podsToDelete []*v1.Pod, pvcs []*v1.PersistentVolumeClaim) (bool, error) {
-	var deleted bool
+	var modified bool
 	for _, pod := range podsToDelete {
-		r.exp.ExpectScale(clonesetutils.GetControllerKey(cs), expectations.Delete, pod.Name)
-		if err := r.Delete(context.TODO(), pod); err != nil {
-			r.exp.ObserveScale(clonesetutils.GetControllerKey(cs), expectations.Delete, pod.Name)
-			r.recorder.Eventf(cs, v1.EventTypeWarning, "FailedDelete", "failed to delete pod %s: %v", pod.Name, err)
-			return deleted, err
+		if cs.Spec.Lifecycle != nil && lifecycle.IsPodHooked(cs.Spec.Lifecycle.PreDelete, pod) {
+			if patched, err := lifecycle.PatchPodLifecycle(r, pod, appsv1alpha1.LifecycleStatePreparingDelete); err != nil {
+				return false, err
+			} else if patched {
+				klog.V(3).Infof("CloneSet %s scaling patch pod %s lifecycle to PreparingDelete",
+					clonesetutils.GetControllerKey(cs), pod.Name)
+				modified = true
+				clonesetutils.ResourceVersionExpectations.Expect(pod)
+			}
+			continue
 		}
-		deleted = true
+
+		clonesetutils.ScaleExpectations.ExpectScale(clonesetutils.GetControllerKey(cs), expectations.Delete, pod.Name)
+		if err := r.Delete(context.TODO(), pod); err != nil {
+			clonesetutils.ScaleExpectations.ObserveScale(clonesetutils.GetControllerKey(cs), expectations.Delete, pod.Name)
+			r.recorder.Eventf(cs, v1.EventTypeWarning, "FailedDelete", "failed to delete pod %s: %v", pod.Name, err)
+			return modified, err
+		}
+		modified = true
 		r.recorder.Event(cs, v1.EventTypeNormal, "SuccessfulDelete", fmt.Sprintf("succeed to delete pod %s", pod.Name))
 
 		// delete pvcs which have the same instance-id
@@ -202,14 +253,14 @@ func (r *realControl) deletePods(cs *appsv1alpha1.CloneSet, podsToDelete []*v1.P
 				continue
 			}
 
-			r.exp.ExpectScale(clonesetutils.GetControllerKey(cs), expectations.Delete, pvc.Name)
+			clonesetutils.ScaleExpectations.ExpectScale(clonesetutils.GetControllerKey(cs), expectations.Delete, pvc.Name)
 			if err := r.Delete(context.TODO(), pvc); err != nil {
-				r.exp.ObserveScale(clonesetutils.GetControllerKey(cs), expectations.Delete, pvc.Name)
+				clonesetutils.ScaleExpectations.ObserveScale(clonesetutils.GetControllerKey(cs), expectations.Delete, pvc.Name)
 				r.recorder.Eventf(cs, v1.EventTypeWarning, "FailedDelete", "failed to delete pvc %s: %v", pvc.Name, err)
-				return deleted, err
+				return modified, err
 			}
 		}
 	}
 
-	return deleted, nil
+	return modified, nil
 }
